Added fontawesome icons & skill bar animations
