+++ src/components/contact.js
@@ -1,4 +1,8 @@
 import React, { useRef } from "react";
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faLinkedin, faGithub } from '@fortawesome/free-brands-svg-icons'
+import { faEnvelope, faLink } from '@fortawesome/free-solid-svg-icons'
+import './sections.css';
 function ContactComponent(props) {
     const contactRef = useRef(null);
     return (
@@ -8,13 +12,13 @@ function ContactComponent(props) {
             <h4>Say hello, give me feedback, or check out my work!</h4>
             <div className="contact-icons">
                 <a className="link-email" href="mailto:evanf.crane@gmail.com">
-                    <i className="fas fa-envelope fa-3x"></i>
+                    <FontAwesomeIcon icon={faEnvelope} size="3x"></FontAwesomeIcon>
                 </a>
                 <a className="link-linkedin" href="https://www.linkedin.com/in/evan-crane/">
-                    <i className="fab fa-linkedin fa-3x"></i>
+                    <FontAwesomeIcon icon={faLinkedin} size="3x"></FontAwesomeIcon>
                 </a>
                 <a className="link-github" href="https://github.com/EvanCrane/">
-                    <i className="fab fa-github fa-3x"></i>
+                    <FontAwesomeIcon icon={faGithub} size="3x"></FontAwesomeIcon>
                 </a>
             </div>
         </section>
